requirements.txt
# Зависимости для Python launcher (легковесный вариант)
pywebview>=4.0.0

